# feat: add startup and performance benchmarks