- Used file scoped namespaces and simplified `Program.cs`
- Added `PerformanceBenchmark` to measure the speed to create and make the average request.
- `StartupBenchmark` measures Refit api creation and call time.
  - Measures: Service creation time, service creation and first call, just first call.
  - Not sure if its worth having the first call benchmarks as they have contradictory results. Why does `FirstCallComplexRequestAsync` use less memory than repeated calls? I might be missing something.

### Performance
| Method                   | Mean      | Error     | StdDev    | Gen0   | Gen1   | Allocated |
|------------------------- |----------:|----------:|----------:|-------:|-------:|----------:|
| ConstantRouteAsync       |  2.940 us | 0.0587 us | 0.0897 us | 0.7744 | 0.0038 |   7.13 KB |
| DynamicRouteAsync        |  3.569 us | 0.0696 us | 0.0905 us | 0.8163 |      - |   7.52 KB |
| ComplexDynamicRouteAsync |  5.022 us | 0.0953 us | 0.1539 us | 0.9003 | 0.0076 |   8.34 KB |
| ObjectRequestAsync       |  5.556 us | 0.0978 us | 0.0817 us | 1.0071 |      - |   9.28 KB |
| ComplexRequestAsync      | 14.040 us | 0.2503 us | 0.2090 us | 1.7090 |      - |  15.86 KB |

### Startup
| Method                       | Job        | InvocationCount | UnrollFactor | Mean     | Error    | StdDev   | Median   | Gen0   | Gen1   | Allocated |
|----------------------------- |----------- |---------------- |------------- |---------:|---------:|---------:|---------:|-------:|-------:|----------:|
| CreateService                | DefaultJob | Default         | 16           | 49.34 us | 0.980 us | 2.733 us | 48.27 us | 5.2490 | 0.1831 |  48.44 KB |
| ConstantRouteAsync           | DefaultJob | Default         | 16           | 52.29 us | 1.023 us | 1.050 us | 52.36 us | 6.1035 | 0.2441 |  57.16 KB |
| ComplexRequestAsync          | DefaultJob | Default         | 16           | 56.78 us | 1.103 us | 1.354 us | 56.49 us | 6.3477 | 0.2441 |  59.34 KB |
| FirstCallConstantRouteAsync  | Job-SUVBDY | 1               | 1            | 43.62 us | 1.434 us | 4.091 us | 42.05 us |      - |      - |   9.87 KB |
| FirstCallComplexRequestAsync | Job-SUVBDY | 1               | 1            | 63.88 us | 1.262 us | 1.640 us | 63.40 us |      - |      - |  13.07 KB |